Rearrange project directories
blog.settings -> config.settings
blog.core -> blog

setup.py and manage.py have been updated to reflect the change
as well as the imports and DJANGO_SETTINGS_MODULES.

This commit separates the django top-level configuration files
into the 'config' package. This now means that the 'blog' package
can be treated as a pure django app.

As the project grows, more apps can be added to the src/ directory tree