Add asgi support
This commit adds the asgi.py file

dj-static is a wsgi middleware, so to allow static files to continue
working in daphne, switch to whitenoise which is a django middleware
that can work with both asgi and wsgi modes.

The default webserver remains gunicorn, but you can run the webserver
using daphne too.

	$ daphne config.asgi:application